LFGIRL0.0.1 contains just data access objects for Users and Groups and they'll deliver the information that their methods claim they will, it's now on someone else to figure out how servlets work and to build some kind of working front end
LFGIRL0.2.3 is me still working in JSF with XHTML, Primefaces, Spring Framework as a Dependency Injector. The login works as per the original way it did when Taylor designed it. I still don't know much about Sessions and their states but I'll see what I can do. There is now functionality to display the groups you're in when you click on "My groups" as well as the ability to search for groups.