User Profiles, Edit Profiles, hide options, to do list
Basic user profiles and editing capabilities, some of the menu options
are hidden when logged out. All menu items lead to a new page rather
than changing mainpage.
I've also made a To Do txt file for stuff that should be done before the
demo this week. Feel free to add anything I've missed, or to remove
stuff when it is done.