Intragroup messaging version 0.1
Added database table and backing bean to support group messaging.

Feature available via my groups.  

Added member list in group profile, messaging function activated from the buttons within each row. Using overlay panel to enter the message, and can check message within the same page using another overlay panel.  

Messaging table and member list is updated via polling, right now at 20 seconds interval.   

Once message is read, the message read status will change from false to ture, indicate message has been read.  The update to the status is also done through the 20 second polling cycle. 

At this version, delete button is not functional, but will be implemented in the next version.  

Message in the database will be removed if the receiving user is deleted, or if the group of which the message belongs to is disbanded. 

The "Send message" button in individual message display window is not working, will be removed in the next version, if time permits will be converted to replay button.    


For version 0.2

-- enable message delete function
-- hide check message button for non-member
-- improve message display panel
-- display sender username instead of user id in the message display table