A little addition of keyboard shortcuts
Keyboard navigation using the "shift" key across buttons and fields on form pages. 
Select a button on focus with "Enter" or "Spacebar" key. Works for pages that contains fields and buttons.(Homepage, login growl, register growl and Account Creation). 
I however didn't manage to do a keyboard navigation to links like on the side menu, or the Login/Sign In.